requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
prometheus_client
gunicorn
python-dotenv
